// lib/generated/LineDirection.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LineStop } from "./LineStop";

export type LineDirection = { headsign: string, stops: Array<LineStop>, 
/**
 * [lon, lat] pairs, Douglas–Peucker simplified
 */
shape: [number, number][], };